# lib/api-zod/src/generated/types/chatMessageInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * F2 FinHeal — Financial Wellness API
 * OpenAPI spec version: 0.1.0
 */

export interface ChatMessageInput {
  message: string;
  session_id: string;
  user_id: string;
}