# lib/api-zod/src/generated/types/moodAnalysis.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * F2 FinHeal — Financial Wellness API
 * OpenAPI spec version: 0.1.0
 */
import type { MoodDimensions } from "./moodDimensions";

export interface MoodAnalysis {
  /** calm | anxious | stressed */
  primary_emotion?: string;
  label?: string;
  dimensions?: MoodDimensions;
}